pull modular fwd transforms out of libjxl_dec
