+++ lib/lib.gni
@@ -180,6 +180,7 @@ libjxl_dec_sources = [
     "jxl/modular/options.h",
     "jxl/modular/transform/palette.h",
     "jxl/modular/transform/rct.h",
+    "jxl/modular/transform/squeeze.cc",
     "jxl/modular/transform/squeeze.h",
     "jxl/modular/transform/transform.cc",
     "jxl/modular/transform/transform.h",
@@ -297,6 +298,14 @@ libjxl_enc_sources = [
     "jxl/modular/encoding/enc_encoding.h",
     "jxl/modular/encoding/enc_ma.cc",
     "jxl/modular/encoding/enc_ma.h",
+    "jxl/modular/transform/enc_palette.cc",
+    "jxl/modular/transform/enc_palette.h",
+    "jxl/modular/transform/enc_rct.cc",
+    "jxl/modular/transform/enc_rct.h",
+    "jxl/modular/transform/enc_squeeze.cc",
+    "jxl/modular/transform/enc_squeeze.h",
+    "jxl/modular/transform/enc_transform.cc",
+    "jxl/modular/transform/enc_transform.h",
     "jxl/optimize.cc",
     "jxl/optimize.h",
     "jxl/progressive_split.cc",
